requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
PyJWT[crypto]>=2.8.0
motor>=3.3.0
pymongo>=4.6.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
